# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IKINCI KANAL - Decentralized Secure Messenger API
  version: 0.2.0
  description: |
    Decentralized, end-to-end encrypted messaging platform.

    ## Core Principles
    - **Zero Knowledge**: Server never sees plaintext content
    - **Metadata Protection**: Sealed sender, minimal logging
    - **User Control**: One-click data deletion
    - **Decentralized**: Federated server architecture
    - **Disappearing Messages**: Auto-delete support

    ## Security Roadmap
    | Version | Authentication | Notes |
    |---------|---------------|-------|
    | **V1 (MVP)** | Passkey (WebAuthn) + password fallback | Passkey preferred, password for legacy |
    | **V2** | OPAQUE protocol | Password users: server NEVER sees password |
    | **V3** | OPAQUE + Hardware key | YubiKey/Titan key binding |

    **V1 Auth Strategy:**
    - **Primary**: Passkey (WebAuthn) - biometric/device-bound, phishing-resistant
    - **Fallback**: Password (Argon2id) - for devices without passkey support
    - **Recommendation**: Prompt users to add passkey even if registered with password

    **Why Passkey in V1?**
    - Mobile: Face ID / Touch ID = zero friction + maximum security
    - Desktop: Platform authenticator or security key
    - No password = no password to steal, leak, or phish

    **Why OPAQUE in V2?**
    For password fallback users: OPAQUE ensures server NEVER sees the password,
    only stores a "password file" useless without client computation.

...
    ```
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /messages/send:
    post:
      tags: [Messages]
      summary: Send encrypted message
      description: |
        Sends an encrypted message. Supports:
        - **Sealed Sender**: Hide sender identity from server
        - **Disappearing**: Auto-delete after read
        - **Federated**: Cross-server delivery
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "202":
          description: Message accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Payload too large (max 256KB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /messages/send-sealed:
    post:
      tags: [Messages]
      summary: Send message with sealed sender (anonymous)
      security: []
      description: |
        **SEALED SENDER**: Server cannot see who sent the message.

        How it works:
        1. Client encrypts sender identity inside the message payload
        2. Client uses a temporary "delivery token" (obtained earlier)
        3. Server only sees: delivery token + recipient + ciphertext
        4. Recipient decrypts to reveal sender identity

        Requires pre-obtained delivery token from recipient.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SealedMessageRequest"
      responses:
        "202":
          description: Sealed message accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "403":
          description: Invalid delivery token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /messages/pending:
    get:
      tags: [Messages]
      summary: Fetch pending messages
      description: |
        Long-poll endpoint for receiving messages.
        Supports waitSeconds for real-time experience without WebSocket.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: waitSeconds
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 30
            default: 0
          description: Long-poll timeout (0 = immediate return)
      responses:
        "200":
          description: Pending messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingMessagesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages/ack:
    post:
      tags: [Messages]
      summary: Acknowledge received messages
      description: |
        Marks messages as delivered, server deletes them from queue.

        **Metadata Privacy Warning:**
        ACKs reveal timing information ("user came online at X").

        **Privacy modes (client-side behavior guidance):**
        - `immediate`: ACK right after decrypt (default, reveals online status)
        - `batched`: Client collects ACKs, sends in random-delayed batches
        - `stealth`: ACK with random 1-60 min delay (max privacy)

        Server doesn't enforce these - it's client UX guidance.
        The `ackMode` field documents client intent for analytics/debugging.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckMessagesRequest"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  ackedCount:
                    type: integer
                required: [ok, ackedCount]
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User authentication token

    federationAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        Server-to-server authentication via Ed25519 request signing.

        **Required Headers:**
        - `X-Server-Origin`: Source server URL (e.g., https://serverA.com)
        - `X-Server-Key-Id`: Public key identifier for rotation (e.g., key-2025-01)
        - `X-Timestamp`: ISO 8601 timestamp (max ±5 min skew)
        - `X-Nonce`: Random 16+ char string (replay protection)
        - `X-Signature`: Base64 Ed25519 signature

        **Signature Computation:**
        ```
        canonical = HTTP_METHOD + "\n"
                  + REQUEST_PATH + "\n"
                  + X-Timestamp + "\n"
                  + X-Nonce + "\n"
                  + SHA256_HEX(request_body)

        signature = Base64(Ed25519.sign(canonical, server_private_key))
        ```

        **Verification:**
        1. Check X-Timestamp within ±5 minutes
        2. Check X-Nonce not seen from this origin in last 5 min (replay)
        3. Fetch sender's public key from `/.well-known/ikincikanal-server`
        4. Verify signature matches canonical request
        5. Cache (origin, nonce) for 5 min

        **Key Rotation:**
        - Rotate keys monthly (recommended)
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: Client-generated key to prevent duplicate sends
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UNAUTHORIZED
            message: Missing or invalid token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: NOT_FOUND
            message: Resource not found
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: RATE_LIMITED
            message: Too many requests
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
...
    SendMessageRequest:
      type: object
      properties:
        toUserId:
          type: string
          format: uuid
        toDeviceId:
          type: string
          format: uuid
          nullable: true
          description: Specific device or null for all devices
        clientMessageId:
          type: string
          format: uuid
        ciphertext:
          type: string
          description: Base64 encrypted message (max 256KB)
          maxLength: 350000
        envelopeType:
          type: string
          enum: [prekey_message, signal_message]
        sentAt:
          type: string
          format: date-time
        disappearingSeconds:
          type: integer
          nullable: true
          minimum: 5
          maximum: 604800
          description: |
            **CLIENT-SIDE** auto-delete timer (NOT server-side).

            - Triggered when: Recipient DECRYPTS and READS the message
            - Timer starts: On first read
            - Deleted by: CLIENT app from local storage
            - Server knows: NOTHING about read status or deletion

            This is different from `expiresAt` (server TTL):
            - `expiresAt`: Server deletes UNDELIVERED messages after 7 days
            - `disappearingSeconds`: Client deletes READ messages after N seconds

            Values: 5 sec - 7 days (604800 sec), null = never auto-delete
        urgent:
          type: boolean
          default: false
          description: Trigger push notification even in DND
      required: [toUserId, clientMessageId, ciphertext, envelopeType, sentAt]
    SealedMessageRequest:
      type: object
      description: |
        Sealed sender message - server cannot see who sent it.
        Sender identity is encrypted inside ciphertext.
      properties:
        deliveryToken:
          type: string
          description: Token obtained from recipient
        toUserId:
          type: string
          format: uuid
        toDeviceId:
          type: string
          format: uuid
          nullable: true
        clientMessageId:
          type: string
          format: uuid
        sealedCiphertext:
          type: string
          description: |
            Base64 sealed ciphertext containing:
            - Sender identity (encrypted)
            - Actual message ciphertext
          maxLength: 360000
        sentAt:
          type: string
          format: date-time
      required: [deliveryToken, toUserId, clientMessageId, sealedCiphertext, sentAt]
    SendMessageResponse:
      type: object
      properties:
        serverMessageId:
          type: string
          format: uuid
        receivedAt:
          type: string
          format: date-time
        queued:
          type: boolean
      required: [serverMessageId, receivedAt, queued]
    PendingMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/PendingMessage"
        serverTime:
          type: string
          format: date-time
        moreAvailable:
          type: boolean
          description: True if more messages waiting
      required: [messages, serverTime, moreAvailable]
    PendingMessage:
      type: object
      properties:
        serverMessageId:
          type: string
          format: uuid
        fromUserId:
          type: string
          format: uuid
          nullable: true
          description: null for sealed sender messages
        fromDeviceId:
          type: string
          format: uuid
          nullable: true
        clientMessageId:
          type: string
        ciphertext:
          type: string
        envelopeType:
          type: string
          enum: [prekey_message, signal_message, sealed_sender]
        receivedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          description: When server will auto-delete undelivered message
        disappearingSeconds:
          type: integer
          nullable: true
          description: Auto-delete timer after read (if set by sender)
      required: [serverMessageId, clientMessageId, ciphertext, envelopeType, receivedAt, expiresAt]
    AckMessagesRequest:
      type: object
      properties:
        serverMessageIds:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: string
            format: uuid
        ackMode:
          type: string
          enum: [immediate, batched, stealth]
          default: immediate
          description: |
            Client privacy behavior (informational, not enforced):
            - immediate: ACK right after reading
            - batched: Collected and sent in batches with random delay
            - stealth: Random 1-60 min delay before ACK

            **Privacy note:** Server does NOT persist ackMode. Used only for
            ephemeral debugging/metrics during request processing, then discarded.
      required: [serverMessageIds]
...